pyproject.toml
[build-system]
requires = ["setuptools",
            "setuptools_scm",
            "wheel"]
build-backend = 'setuptools.build_meta'

[project]
name = "synthesizAR"
dynamic = ["version"]
authors = [
  {name="Will Barnes", email="will.t.barnes@gmail.com"}
]
license = {file="LICENSE.rst"}
description = "A Python package for forward modeling optically-thin emission from field-aligned hydrodynamic simulations of astrophysical plasmas."
readme = {file="README.md", content-type = "text/markdown"}
requires-python = ">=3.10"
dependencies = [
    "scipy",
    "matplotlib",
    "sunpy[all]>=6.0",
    "zarr",
    "asdf",
    "ndcube>=2.0",
]

[project.urls]
Repository = "https://github.com/wtbarnes/synthesizAR"
Homepage = "https://github.com/wtbarnes/synthesizAR"
Documentation = "https://synthesizar.readthedocs.io"

[project.optional-dependencies]
all = ["synthesizAR[aia,atomic,ebtel,hydrad,parallel,xrt]"]
aia = ["aiapy"]
atomic = [
  "plasmapy",
  "fiasco",
]
ebtel = ["ebtelplusplus"]
hydrad = ["pydrad@git+https://github.com/rice-solar-physics/pydrad@main"]
parallel =[
  "dask[complete]",
  "distributed",
]
xrt = ["xrtpy"]
test =[
  "synthesizAR[all]",
  "tox",
  "flake8",
  "pytest",
  "pytest-astropy",
  "pytest-cov",
]
docs = [
  "synthesizAR[all]",
  "pydata-sphinx-theme",
  "sphinx-automodapi",
  "sphinx-design",
  "sphinx-gallery",
  "astroquery",
  "sphinx",
  "sphinx-astropy",
  "sphinxcontrib-bibtex",
]
dev = ["synthesizAR[test,docs]"]

[tool.setuptools]
packages = ["synthesizAR"]

[tool.setuptools_scm]
write_to = "synthesizAR/version.py"

[tool.pytest.ini_options]
testpaths = [
  "synthesizAR",
  "docs",
]
norecursedirs = [
  "docs/_build",
  "docs/generated",
]
doctest_plus = "enabled"
text_file_format = "rst"
addopts = "--doctest-rst"
remote_data_strict = "True"
filterwarnings = [
  "error",
  "ignore:The unit 'G' has been deprecated in the VOUnit standard.*:astropy.units.core.UnitsWarning",
  "ignore:'cgi' is deprecated and slated for removal in Python 3.13:DeprecationWarning",
]

[tool.coverage.run]
branch = true
omit = [
  "synthesizAR/conftest.py",
  "synthesizAR/tests/*",
  "synthesizAR/*/tests/*",
  "synthesizAR/version*",
  "*/synthesizAR/conftest.py",
  "*/synthesizAR/tests/*",
  "*/synthesizAR/*/tests/*",
  "*/synthesizAR/extern/*",
  "*/synthesizAR/version*",
]

[tool.coverage.report]
exclude_lines =[
  # Have to re-enable the standard pragma
  "pragma: no cover",
  # Don't complain about packages we have installed
  "except ImportError",
  # Don't complain if tests don't hit assertions
  "raise AssertionError",
  "raise NotImplementedError",
  # Ignore branches that don't pertain to this version of Python
  "pragma: py{ignore_python_version}",
  # Don't complain about IPython completion helper
  "def _ipython_key_completions_",
]

[tool.isort]
line_length = 88
wrap_length = 80
sections = ["FUTURE", "STDLIB", "FIRSTPARTY", "LOCALFOLDER"]
known_first_party = ["synthesizAR", ]
default_section = "STDLIB"
multi_line_output = 3
use_parentheses = true
include_trailing_comma = true
force_alphabetical_sort_within_sections = true
honor_noqa = true
lines_between_types = 1

[tool.codespell]
skip = "*.fts,*.fits,venv,*.pro,*.bib,*.asdf,*.json"
ignore-words-list = "te,emiss"

[tool.ruff]
target-version = "py310"
line-length = 110
exclude=[
  ".git,",
  "__pycache__",
  "build",
  "synthesizAR/version.py",
]
show-fixes = true
output-format = "full"

lint.select = [
  "E",
  "F",
  "W",
  "UP",
  "PT",
  #"RET",
  #"TID",
]
lint.extend-ignore = [
  # pycodestyle (E, W)
  "E501",  # LineTooLong                             # TODO! fix
  "E741",  # Ambiguous variable name

  # pytest (PT)
  "PT001",  # Always use pytest.fixture()
  "PT004",  # Fixtures which don't return anything should have leading _
  "PT007",  # Parametrize should be lists of tuples  # TODO! fix
  "PT011",  # Too broad exception assert             # TODO! fix
  "PT023",  # Always use () on pytest decorators
]

[tool.ruff.lint.per-file-ignores]
# Part of configuration, not a package.
"setup.py" = ["INP001"]
"conftest.py" = ["INP001"]
# implicit-namespace-package. The examples are not a package.
"docs/*.py" = ["INP001"]
# Module level imports do not need to be at the top of a file here
"docs/conf.py" = ["E402"]
"__init__.py" = ["E402", "F401", "F403"]
"test_*.py" = ["B011", "D", "E402", "S101"]

[tool.ruff.lint.pydocstyle]
convention = "numpy"
